requirements.txt
openai
python-dotenv
aiogram          # для Telegram-бота
whisper          # OpenAI-клиент для ASR